// products-client-app/src/app/models/Product.ts
export interface Product {
  id: string;
  name: string;
  shortName: string;
  description: string;
  shortDescription: string;
  price: number;
  category: string;
  image: string;
  height: number;
  width: number;
  depth: number;
  weight: number;
  quantity: number;
  rating: number;
  color: string;
  shipping: string;
  date: string;
}